+++ atomicshop/permissions.py
@@ -29,43 +29,16 @@ def is_admin() -> bool:
     return result
 
 
-def request_sudo_on_ubuntu_by_python():
+def get_ubuntu_sudo_executer_username() -> str:
     """
-    The function tries to request sudo on Ubuntu for the user to enter the password, by executing python executable
-    with sudo. This is better approach than using the bash script to request sudo (request_sudo_on_ubuntu_by_bash)
-    because of the permission issues that can occur.
-
-    :return:
+    Function gets the username of the user who executed the script with sudo.
+    :return: str, username.
     """
 
-    try:
-        # Attempt to re-execute the script using sudo
-        subprocess.check_call(['sudo', 'python3'] + sys.argv)
-    except subprocess.CalledProcessError:
-        # Handle the error in case sudo command fails (e.g., wrong password)
-        print_api("Failed to gain sudo access. Please try again.", color='red')
-        sys.exit(1)
-
-
-def request_sudo_on_ubuntu_by_bash():
-    """
-    The function tries to request sudo on Ubuntu for the user to enter the password, by executing appropriate
-    bash commands.
-    Better approach is to use 'request_sudo_on_ubuntu_by_python'.
-
-    :return:
-    """
-
-    script = """
-    if [ "$EUID" -ne 0 ]; then
-      echo "This script requires root privileges. Please enter your password for sudo access."
-      sudo -v
-      while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
-    fi
-    # Your bash commands that require sudo here
-    """
-
-    process.execute_script(script, shell=True)
+    if 'SUDO_USER' in os.environ:
+        return os.environ['SUDO_USER']
+    else:
+        return ''
 
 
 def set_executable_permission(file_path: str):